Split DepContextJsonReader into common, newtonsoft, and netstandard.
